requirements.txt
ipython

pytest

python-dotenv

requests
